Refactoring Download
* Shoes3/shoes3#333
